+++ src/models/feedback.ts
@@ -4,11 +4,8 @@ import sequelize from '@/config/db';
 // Define interface for Feedback attributes
 interface FeedbackAttributes {
   id?: number;
-  canteen: number | null;
   food: number | null;
-  system: number | null;
   content: string | null;
-  userId?: string;
   dishId?: number;
   createdAt?: Date;
   updatedAt?: Date;
@@ -20,11 +17,8 @@ interface FeedbackCreationAttributes extends Optional<FeedbackAttributes, 'id' |
 // Define Feedback model class
 class Feedback extends Model<FeedbackAttributes, FeedbackCreationAttributes> implements FeedbackAttributes {
   public id!: number;
-  public canteen!: number | null;
   public food!: number | null;
-  public system!: number | null;
   public content!: string | null;
-  public userId!: string | undefined;
   public dishId!: number | undefined;
   
   // Timestamps
@@ -39,15 +33,7 @@ Feedback.init(
       primaryKey: true,
       autoIncrement: true,
     },
-    canteen: {
-      type: DataTypes.INTEGER,
-      allowNull: true
-    },
     food: {
-      type: DataTypes.INTEGER,
-      allowNull: true
-    },
-    system: {
       type: DataTypes.INTEGER,
       allowNull: true
     },
@@ -58,10 +44,6 @@ Feedback.init(
         len: [0, 250]
       }
 
-    },
-    userId: {
-      type: DataTypes.UUID,
-      allowNull: true
     },
   }, 
   {
